Refactor: move the generation scripts to a separate file
